+++ libvast/vast/uuid.hpp
@@ -21,6 +21,7 @@
 #include <caf/error.hpp>
 #include <caf/expected.hpp>
 #include <caf/meta/hex_formatted.hpp>
+#include <caf/meta/type_name.hpp>
 
 #include <array>
 
@@ -70,7 +71,8 @@ public:
 
   template <class Inspector>
   friend auto inspect(Inspector& f, uuid& x) {
-    return f(caf::meta::hex_formatted(), x.id_);
+    return f(caf::meta::type_name("vast.uuid"), caf::meta::hex_formatted(),
+             x.id_);
   }
 
 private: